Updated the matrix in matrixTranspose file

+++ matrixTranspose.cpp
@@ -1,9 +1,10 @@
 #include "matrixTranspose.h"
 
-void MatrixTranspose(float matrix[5][5]) {
+void MatrixTranspose(double matrix[5][5]) {
+    // Transpose the matrix
     for (int i = 0; i < 5; i++) {
         for (int j = i; j < 5; j++) {
-            float temp = matrix[i][j];
+            double temp = matrix[i][j];
             matrix[i][j] = matrix[j][i];
             matrix[j][i] = temp;
         }
