+++ matrixTranspose.h
@@ -1,6 +1,6 @@
 #ifndef MATRIXTRANSPOSE_H
 #define MATRIXTRANSPOSE_H
 
-void MatrixTranspose(float matrix[5][5]);
+void MatrixTranspose(double matrix[5][5]);
 
 #endif
